add linux support to desktop patch script
On linux, the user must pass in the path to the resources directory
as an argument to the script. Also adjusted a couple of other items
like "sudo env ..." to work properly in linux.

Added a note to the README to ensure that the user knows to run
Slack with the dev mode on.